refactor(extractor): extracting filter functions and transforming object to JSON
