+++ src/scripts/content/extractor/turndownConfPerplexity.json
@@ -1,0 +1,15 @@
+{
+  "init": {
+    "blankReplacement": "getBlankReplacement"
+  },
+  "rules": {
+    "preserveLineBreaksInPre": {
+      "filter": "filter_PreserveLineBreaksInPre_Perplexity",
+      "replacement": "replacement_PreserveLineBreaksInPre_Perplexity"
+    },
+    "formatCitationsInAnswer": {
+      "filter": "filter_formatCitationsInAnswer_Perplexity",
+      "replacement": "replacement_formatCitationsInAnswer_Perplexity"
+    }
+  }
+}
